+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added `zero-shot-classification` pipeline
 - Added `zero-shot-image-classification` pipeline
+- Added `zero-shot-object-detection` pipeline
 - Added `fill-mask` pipeline
 - Added `image-classification` pipeline
 - Added `image-feature-extraction` pipeline